hmr: keep provider api
